+++ phpstan.neon
@@ -3,7 +3,7 @@ parameters:
 
     excludes_analyse:
         - %rootDir%/../../../scripts/Phalcon/Generator/*
-        - %rootDir%/../../../scripts/Phalcon/Commands/DotPhalconMissingException.php
+        - %rootDir%/../../../scripts/Phalcon/Exception/*
     ignoreErrors:
         - '#Call to an undefined method object::toArray().#'
         - '#supplied for foreach, only iterables are supported#'
