+++ scripts/Phalcon/Commands/Builtin/Webtools.php
@@ -4,7 +4,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
+  | Copyright (c) 2011-present Phalcon Team (https://www.phalconphp.com)   |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file LICENSE.txt.                             |
@@ -20,10 +20,10 @@
 
 namespace Phalcon\Commands\Builtin;
 
+use Phalcon\Web\Tools;
 use Phalcon\Script\Color;
 use Phalcon\Commands\Command;
-use Phalcon\Web\Tools;
-use Phalcon\Commands\CommandsException;
+use Phalcon\Exception\Commands\CommandsException;
 
 /**
  * Webtools Command
