+++ scripts/Phalcon/Exception/Commands/CommandsException.php
@@ -4,7 +4,7 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
+  | Copyright (c) 2011-present Phalcon Team (https://www.phalconphp.com)   |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
   | with this package in the file LICENSE.txt.                             |
@@ -18,15 +18,15 @@
   +------------------------------------------------------------------------+
 */
 
-namespace Phalcon\Commands;
-
-use Phalcon\Exception;
+namespace Phalcon\Exception\Commands;
 
 /**
+ * Phalcon\Exception\Commands\CommandsException
+ *
  * Commands Exception
  *
- * @package Phalcon\Commands
+ * @package Phalcon\Exception\Commands
  */
-class CommandsException extends Exception
+class CommandsException extends \InvalidArgumentException
 {
 }
